chore: modularise script
