+++ scraper/scraper.py
@@ -1,5 +1,7 @@
 import cloudscraper
 import json
+import scraper_utils
+from scraper_source import slugs
 
 scraper = cloudscraper.create_scraper(
     browser={
@@ -8,39 +10,6 @@ scraper = cloudscraper.create_scraper(
         'desktop': True 
     }
 )
-
-def filter_typename(dict):
-  return dict["__typename"] == "AssetQuantityType"
-
-def filter_images(dict):
-  return dict["__typename"] == "AssetType"
-
-def filter_metadata(dict):
-  if "tokenId" in dict:
-    return {
-      'id': int(dict["tokenId"]),
-      'name': dict["name"],
-      'image': dict["displayImageUrl"]
-    }
-  else:
-    return
-
-def filter_quantityInEth_exists(dict):
-  if "quantityInEth" in dict:
-    return True
-  else:
-    return False
-
-def get_floor_price_in_eth(dict):
-  return float(dict["quantity"]) / 1000000000000000000
-
-def format_result(pair):
-  return {
-    "price": pair[0],
-    "id": pair[1]["id"],
-    "name": pair[1]["name"],
-    "image": pair[1]["image"]
-  }
 
 def get_floor_prices(slug):
   url = "https://opensea.io/collection/{}?search[sortAscending]=true&search[sortBy]=PRICE&search[toggles][0]=BUY_NOW".format(slug);
@@ -51,28 +20,18 @@ def get_floor_prices(slug):
   data_values = data["records"].values() # get all values type...
   data_list = [*data_values]
   # get metadata of images
-  images = list(filter(filter_images, data_list))
-  images = list(map(filter_metadata, images))
+  images = list(filter(scraper_utils.filter_images, data_list))
+  images = list(map(scraper_utils.filter_metadata, images))
   images = list(filter(lambda x: x is not None, images))
   # get prices of images
-  data_list = list(filter(filter_typename, data_list))
-  prices_data_list = list(filter(filter_quantityInEth_exists, data_list))
-  prices_data_list = list(map(get_floor_price_in_eth, prices_data_list))
+  data_list = list(filter(scraper_utils.filter_typename, data_list))
+  prices_data_list = list(filter(scraper_utils.filter_quantityInEth_exists, data_list))
+  prices_data_list = list(map(scraper_utils.get_floor_price_in_eth, prices_data_list))
   prices_data_list = sorted(prices_data_list)
-  
+  # format results
   result = list(zip(prices_data_list, images))
-  result = list(map(format_result, result))
+  result = list(map(scraper_utils.format_result, result))
   return result
-
-def write_data_to_file(filename, data):
-  with open(filename, 'w') as f:
-    json.dump(json.dumps(data, ensure_ascii=False, indent=4), f)
-  f.close()
-
-def write_json_to_file(filename, json_string):
-  with open(filename, 'w') as f:
-    f.write(json_string)
-  f.close()
 
 # Structure (priceInEth: float, { id: int, name: str, image: str })
 def scrape_slug(slug: str):
@@ -80,17 +39,27 @@ def scrape_slug(slug: str):
   try:
     data = get_floor_prices(slug)
     print("Total {} NFTS scraped".format(len(data)))
+    scraper_utils.write_json_to_file('scraped-collections/{}.json'.format(slug), data)
     for obj in data:
       print(obj)
-  except:
-    print("@@@@@@@@@@@@@@ ERROR ERROR ERROR @@@@@@@@@@@@@@@")
+  except Exception as e:
+    print("@@@@@@@@@@@@@@ ERROR ERROR ERROR @@@@@@@@@@@@@@@ ")
+    print(e)
+    print("FAILURE: Unable to scrape for {} collection".format(slug))
   print("=================== Ended ========================")
 
-from scraper_source import slugs
-# TEST
-test = input('Scrape all? (Y/N)')
-if test == 'Y' or test == 'y' or test == 'yes':
-  for collection in slugs:
-    scrape_slug(collection)
+def scrape_all():
+  test = input('Scrape all? (Y/N)')
+  if test == 'Y' or test == 'y' or test == 'yes':
+    for collection in slugs:
+      scrape_slug(collection)
     
-scrape_slug(slugs[2])
+# TEST
+def main():
+  # scrape_slug(slugs[2])
+  scrape_all()
+
+if __name__ == "__main__":
+    main()
+
+
